Move SignalWorker out into a separate file
