+++ swift/Workflow/Sources/Worker.swift
@@ -44,28 +44,3 @@ extension Worker where Self: Equatable {
     }
 
 }
-
-/// A Worker that wraps a `Signal`
-public struct SignalWorker<Key: Equatable, Value>: Worker {
-    let key: Key
-    let signal: Signal<Value, Never>
-
-    public init(key: Key, signal: Signal<Value, Never>) {
-        self.key = key
-        self.signal = signal
-    }
-
-    public func run() -> SignalProducer<Value, Never> {
-        return SignalProducer(signal)
-    }
-
-    public func isEquivalent(to otherWorker: SignalWorker) -> Bool {
-        return key == otherWorker.key
-    }
-}
-
-extension Signal where Error == Never {
-    public func asWorker<Key: Equatable>(key: Key) -> SignalWorker<Key, Value> {
-        return SignalWorker(key: key, signal: self)
-    }
-}
